fix: zeroize recovery keys, fix gen_id race, Content-Length truncation
- gen_id: SeqCst ordering prevents duplicate IDs under concurrency
- http.rs: Content-Length compared as u64, not truncated via `as usize`
- did.rs: generate_recovery_keypair returns Zeroizing<[u8; 32]>,
  mnemonic_to_private_key zeroizes entropy and returns Zeroizing
- did.rs: hex_encode uses pre-allocated buffer instead of per-byte format!
- sanitize.rs: html_to_text uses cached LazyLock Builder
- media.rs: image limits include max_image_width/height for GIF defense

+++ src/http.rs
@@ -5,7 +5,7 @@ use anyhow::{bail, Context, Result};
 pub async fn read_body_limited(mut resp: reqwest::Response, max_bytes: usize) -> Result<Vec<u8>> {
     // Fast-reject via Content-Length if available (advisory but cheap)
     if let Some(len) = resp.content_length() {
-        if len as usize > max_bytes {
+        if len > max_bytes as u64 {
             bail!("response Content-Length {len} exceeds {max_bytes} byte limit");
         }
     }
